requirements.txt
Flask
transformers
torch
pdfplumber
openai
python-dotenv
